refactor: P2 server split 11/N — extract chat route domain (final)
Move all /api/chat* routes out of server.ts into server/routes/chat.ts as
handleChatRoutes: non-streaming POST /api/chat, question-response, scoped
abort, status, SSE event replay, and the SSE streaming turn (the largest
block, including the heartbeat, workspace-change monitor, and the
persistence-confirmation finalize path).

Chat-only helpers move with it: piEventToSseEvent +
toolCallStreamEventFromAssistantEvent, persistInlineImages +
mimeTypeToExtension + prependImagePathsHint, and the workspace-change
monitor family. readSessionBaseline/confirmTurnPersistence move too, now
taking parseSessionFile/sessionRevision as explicit parameters (those two
stay in server.ts — maybeAutoGenerateTopic and the sessions ctx still use
them — and are injected via ctx alongside the other session-metadata
helpers). The unused ImageContent import is dropped.

server.ts: 2267 → 1608 lines (5295 at P2 start, -70%). What remains:
bootstrap, channel wiring/hot-reload, queue helpers, skill installers,
static/SPA serving, the terminal WebSocket upgrade, and eleven one-line
route delegations.

Smoke coverage: /api/chat 400 without prompt, /api/chat/stream 400 without
fields, status found:false, scoped abort 404.
